Extend filesystem fuzzing

+++ include/fuzzing/testers/serialize/filesystem.hpp
@@ -3,6 +3,7 @@
 #include <fuzzing/datasource/datasource.hpp>
 #include <fuzzing/datasource/id.hpp>
 #include <fuzzing/generators/filesystem.hpp>
+#include <fuzzing/exception.hpp>
 #include <stdlib.h>
 
 namespace fuzzing {
@@ -23,22 +24,34 @@ class FilesystemTester {
 
         bool Run(void) {
             if ( fs.Write() == false ) {
+                /* Write can fail due to overlong/invalid paths.
+                 * Not a critical error
+                 */
                 return false;
             }
 
             if ( fs.Verify() == false ) {
+                throw exception::TargetException(
+                    "Writing succeeded, but verifying the written data failed"
+                );
                 return false;
             }
 
             if ( transform() == false ) {
-                return true;
+                return false;
             }
 
             if ( fs.Verify() == false ) {
+                throw exception::TargetException(
+                    "Verifying data failed after transformation"
+                );
                 return false;
             }
 
             if ( fs.Remove() == false ) {
+                throw exception::TargetException(
+                    "Removing the data failed"
+                );
                 return false;
             }
 
@@ -46,42 +59,79 @@ class FilesystemTester {
         }
 };
 
-class TarTester : public fuzzing::testers::filesystem::FilesystemTester {
+class ArchiverTester : public FilesystemTester {
+    protected:
+        virtual bool pack(const std::string infile, const std::string outfile) = 0;
+        virtual bool unpack(const std::string infile) = 0;
     private:
         bool transform(void) {
-            {
-                static const auto createTarCmd = std::string("tar cf archive.tar " + fsRootPath + "/");
-                if ( system(createTarCmd.c_str()) != 0 ) {
-                    abort();
-                }
+            if ( pack(fsRootPath + "/", "archive") == false ) {
+                return false;
             }
 
             {
                 static const auto rmdirCmd = std::string("rm -rf " + fsRootPath + "/");
                 if ( system(rmdirCmd.c_str()) != 0 ) {
-                    abort();
+                    throw exception::TargetException(
+                        "Removing the working directory failed"
+                    );
                 }
             }
 
-            {
-                static const auto extractCmd = std::string("tar xf archive.tar");
-                if ( system(extractCmd.c_str()) != 0 ) {
-                    abort();
-                }
+            if ( unpack("archive") == false ) {
+                throw exception::TargetException(
+                    "Tar cannot process its own data"
+                );
+                return false;
             }
 
             {
                 static const auto rmTarCmd = std::string("rm -rf archive.tar");
                 if ( system(rmTarCmd.c_str()) != 0 ) {
-                    abort();
+                    throw exception::TargetException(
+                        "Removing the archive failed"
+                    );
                 }
             }
 
             return true;
         }
     public:
-        TarTester(fuzzing::datasource::Datasource& ds, const std::string fsRootPath) :
-            fuzzing::testers::filesystem::FilesystemTester(ds, fsRootPath)
+        ArchiverTester(fuzzing::datasource::Datasource& ds, const std::string fsRootPath) :
+            FilesystemTester(ds, fsRootPath)
+        { }
+};
+
+class TarTester : public ArchiverTester {
+    private:
+        const std::string tarCmd;
+
+        bool pack(const std::string infile, const std::string outfile) override {
+            const auto cmd = std::string(tarCmd + " cf " + outfile + " " + infile);
+
+            if ( system(cmd.c_str()) != 0 ) {
+                return false;
+            }
+
+            return true;
+        }
+
+        bool unpack(const std::string infile) override {
+            const auto cmd = std::string(tarCmd + " xf " + infile);
+
+            if ( system(cmd.c_str()) != 0 ) {
+                return false;
+            }
+
+            return true;
+        }
+    public:
+        TarTester(
+                fuzzing::datasource::Datasource& ds,
+                const std::string fsRootPath,
+                const std::string tarCmd) :
+            ArchiverTester(ds, fsRootPath),
+            tarCmd(tarCmd.empty() ? "tar" : tarCmd)
         { }
 };
